Flattening the curve
It is a well known fact that the virus spread takes an exponential distribution. It is a well established fact that social distancing can flatten the curve (The bell curve). It is necessary to flatten the curve so much so that healthcare infrastructure in the country can manage the peak number of patients at any given time.

The dataset contains the details of number of beds and hospitals (statewise distribution can be found here). The database also contains individual patient data. (for real time data access here).

Updates : Indian Railways is allotting train compartments as isolation wards. Database of Indian railways can be found here